Changed mana costs to be of type float Changed most modifers to cost 1.25 mana Changed some components mana cost Increased the default mana points per level to 200

+++ Framework/Interfaces/Spells/ISpell.cs
@@ -76,7 +76,7 @@ namespace ArsVenefici.Framework.Interfaces.Spells
         /// </summary>
         /// <param name="caster">The player that casts this spell.</param>
         /// <returns>The amount of mana this spell costs.</returns>
-        int Mana();
+        float Mana();
 
         /// <returns>The list of spell parts in this spell.</returns>
         List<ISpellPart> Parts();
