+++ Framework/Spell/Components/CreateWater.cs
@@ -73,7 +73,7 @@ namespace ArsVenefici.Framework.Spell.Components
             }
         }
 
-        public override int ManaCost()
+        public override float ManaCost()
         {
             return 5;
         }
