+++ Framework/Spell/Components/Dig.cs
@@ -264,7 +264,7 @@ namespace ArsVenefici.Framework.Spell.Components
             return false;
         }
 
-        public override int ManaCost()
+        public override float ManaCost()
         {
             return 5;
         }
